requirements.txt
ynab>=1.9.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
